// src/@fuse/types/fuse-navigation.ts
export interface FuseNavigationItem {
  id: string;
  title?: string;
  type: 'item' | 'group' | 'collapsable';
  translate?: string;
  icon?: string;
  svgIcon?: string;
  hidden?: boolean;
  url?: string;
  classes?: string;
  exactMatch?: boolean;
  externalUrl?: boolean;
  openInNewTab?: boolean;
  function?: any;
  disabled?: boolean;
  badge?: {
    title?: string;
    translate?: string;
    bg?: string;
    fg?: string;
  };
  children?: FuseNavigationItem[];
}

export interface FuseNavigation extends FuseNavigationItem {
  children?: FuseNavigationItem[];
}